Initialitation Error solved and RecyclingZoneService Test using mockito

+++ src/main/java/com/example/proyectoFinal1/dto/RecyclingZoneDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class RecyclingZoneDTO {
-    private Long recyclingZoneId;
+    private int recyclingZoneId;
     private String RecyclingZoneName;
     private String RecyclingZoneAdress;
